requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
ajax==3.0.0
asgiref==3.8.1
AsyncKandinsky==2.0.1    
attrs==23.2.0
blinker==1.7.0
certifi==2024.2.2        
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
decorator==5.1.1
Flask==3.0.3
frozenlist==1.4.1
gunicorn==22.0.0
helpers==0.2.0
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.3
MarkupSafe==2.1.5
multidict==6.0.5
node==1.2.1
odict==1.9.0
packaging==24.0
pillow==10.3.0
plumber==1.7
requests==2.31.0
setuptools==69.5.1
sqlparse==0.5.0
tzdata==2024.1
urllib3==2.2.1
websockets==12.0
Werkzeug==3.0.2
yarl==1.9.4
zope.component==6.0
zope.deferredimport==5.0
zope.deprecation==5.0
zope.event==5.0
zope.hookable==6.0
zope.interface==6.3
zope.lifecycleevent==5.0
zope.proxy==5.2